requirements.txt
python-dotenv==0.10.1
python-telegram-bot==11.1.0
vk-api==11.4.0
requests==2.21.0
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
oauth2client==4.1.3
PyDrive==1.3.1
pandas==0.24.1
